+++ atlas/modules/resource_creator/creators.py
@@ -235,8 +235,7 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
     def update(self):
 
         # Update Specs File
-        self.write_file_to_output(self.swagger_file, self.specs, append_mode=False,
-                                  project_sub_folder=settings.ARTILLERY_LIB_FOLDER)
+        self.write_file_to_output(self.swagger_file, self.specs, append_mode=False)
 
         # Update Resource Mapping File
         auto_resource = {resource: {"def": "# Add your definition here"} for resource in self.new_resources}